+++ packages/theme/src/composables/config/blog.ts
@@ -186,6 +186,10 @@ export function useBlogThemeMode() {
   return inject(configSymbol)!.value?.blog?.blog ?? true
 }
 
+export function useBlogInfoCollapsible() {
+  return inject(configSymbol)!.value?.blog?.home?.blogInfoCollapsible ?? false
+}
+
 export function useArticles() {
   const blogConfig = useConfig()
   const { localeIndex, site } = useData()
@@ -357,7 +361,7 @@ export function useHomeAnalysis() {
 }
 
 export function useAnalyzeTitles(wordCount: Ref<number>, readTime: ComputedRef<number>) {
-  const article = computed(() => useConfig()?.value.blog?.article)
+  const article = useArticleConfig()
 
   const topWordCount = computed(() =>
     replaceValue(article.value?.analyzeTitles?.topWordCount || '字数：{{value}} 个字', wordCount.value)