+++ packages/theme/docs/changelog.md
@@ -1,6 +1,6 @@
 ---
 title: 更新日志
-description: 最近更新（v0.5.3） ⏰ 2024/12/21：pagefind 不检索文章标题
+description: 最近更新（v0.5.6） ⏰ 2025/08/10：系列bug修复
 author: 粥里有勺糖
 top: 3
 tag: 日志
@@ -26,6 +26,27 @@ bun update @sugarat/theme
 bun install vitepress@latest
 ```
 :::
+
+## 0.5.6 (2025/08/10)
+
+### Patch Changes
+
+- feat: 支持移动端信息卡片折叠
+- fix: HMR Vue Warn inject
+- fix: 多标签展示异常
+
+## 0.5.5 (2025/07/03)
+
+### Patch Changes
+
+- feat: 支持按需开启 group-icons 插件
+- fix: 一些 SVG 图标 iOS 不渲染
+- fix: 日期在 iOS 上显示 NaN
+- fix: 作品页样式塌陷
+- fix: 锁定依赖 group-icons 版本
+- Updated dependencies
+  - vitepress-plugin-announcement@0.1.5
+
 ## 0.5.4 (2025/01/25)
 
 ### Patch Changes